Create TIMESTOMPING_EXPLAINED_ON_API_LEVEL.pdf
While working on unpacking a malware, I encountered a challenge that I have yet to overcome as I write this. However, upon successfully executing the malware, I noticed that it had dropped a file onto the disk with a different timestamp. I knew Time Stomping is used but this made me curious to explore the concept on lower level and examining how malware utilizes it to manipulate timestamps and how we can determine the original timestamp set for a file.